+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added *Paramater* option to *Spline Falloff* node.
 - Added *Wrap Paramaters* option to *Evaluate Spline* node.
 - Added *Matrix* output to the *Evaluate Spline* node.
+- Added search tags for Distribute Matrices node.
 
 ### Fixed